added smart traveller spider
