--- PHASE_1/API_SourceCode/scraper/.gitignore
@@ -1,4 +1,0 @@
-/__pycache__
-src/__pycache__
-src/WHODON_scraper/__pycache__
-src/WHODON_scraper/spiders/__pycache__
