+++ PHASE_1/API_SourceCode/scraper/WHO/.gitignore
@@ -1,0 +1,4 @@
+/__pycache__
+src/__pycache__
+src/WHODON_scraper/__pycache__
+src/WHODON_scraper/spiders/__pycache__
