+++ PHASE_1/API_SourceCode/scraper/WHO/config/example.txt
@@ -1,0 +1,33 @@
+once a day, 
+
+middleware calls flask server to scrape data
+
+~an hour later, receive the file using /articles
+
+read json
+
+process each entry in the json
+
+update the database
+    - finding key terms
+    - logging new articles
+    - 
+
+CREATE TABLE articles (
+    id SERIAL NOT NULL,
+    url STRING NOT NULL,
+    headline STRING NOT NULL,
+    main_text STRING NOT NULL,
+    date_of_publication DATETIME/TIMESTAMP NOT NULL,
+    
+    PRIMARY KEY id,
+)
+
+CREATE TABLE reports (
+    id SERIAL PRIMARY KEY,
+    disease STRING[] NOT NULL,
+    syndrome STRING[] NOT NULL,
+    event_date DATETIME/TIMESTAMP NOT NULL,
+    locations STRING[], /*may become INTEGER if we are using geoIDs*/
+    article FOREIGN KEY (id) REFERENCES articles(id)
+)
