--- PHASE_1/API_SourceCode/scraper/scrapy.cfg
@@ -1,13 +1,0 @@
-# Automatically created by: scrapy startproject
-#
-# For more information about the [deploy] section see:
-# https://scrapyd.readthedocs.io/en/latest/deploy.html
-
-[settings]
-default = WHODON_scraper.settings
-
-[deploy]
-url = https://scrapy-test-server.herokuapp.com/
-username = admin
-password = scrapydweb
-project = WHODON_scraper
