--- PHASE_1/API_SourceCode/scraper/src/WHODON_scraper/key_terms.py
@@ -1,29 +1,0 @@
-import json
-
-# converts pasted raw json text into an array
-
-raw = """{"payload":[ { "name": "unknown" },{ "name": "other" },{ "name": "anthrax cutaneous" },{ "name": "anthrax gastrointestinous" },{ "name": "anthrax inhalation" },{ "name": "botulism" },{ "name": "brucellosis" },{ "name": "chikungunya" }, { "name": "cholera" },
-{ "name": "cryptococcosis" },{ "name": "cryptosporidiosis" },{ "name": "crimean-congo haemorrhagic fever" },{ "name": "dengue" },{ "name": "diphteria" },{ "name": "ebola haemorrhagic fever" },{ "name": "ehec (e.coli)" },{ "name": "enterovirus 71 infection" },{ "name": "influenza a/h5n1" },{ "name": "influenza a/h7n9" },{ "name": "influenza a/h9n2" },{ "name": "influenza a/h1n1" },{ "name": "influenza a/h1n2" },{ "name": "influenza a/h3n5" },{ "name": "influenza a/h3n2" },{ "name": "influenza a/h2n2" },{ "name": "hand, foot and mouth disease" },{ "name": "hantavirus" },{ "name": "hepatitis a" },{ "name": "hepatitis b" },{ "name": "hepatitis c" }, { "name": "hepatitis d" },{ "name": "hepatitis e" },{ "name": "histoplasmosis" },{ "name": "hiv/aids" },{ "name": "lassa fever" },{ "name": "malaria" },{ "name": "marburg virus disease" },{ "name": "measles" },{ "name": "mers-cov" },{ "name": "mumps" },{ "name": "nipah virus" },{ "name": "norovirus infection" },{ "name": "pertussis" },{ "name": "plague" },{ "name": "pneumococcus pneumonia" },{ "name": "poliomyelitis" },{ "name": "q fever" },{ "name": "rabies" },{ "name": "rift valley fever" },{ "name": "rotavirus infection" },{ "name": "rubella" },{ "name": "salmonellosis" },{ "name": "sars" },{ "name": "shigellosis" },{ "name": "smallpox" },{ "name": "staphylococcal enterotoxin b" },{ "name": "thypoid fever" },{ "name": "tuberculosis" },{ "name": "tularemia" },{ "name": "vaccinia and cowpox" },{ "name": "varicella" },{ "name": "west nile virus" },{ "name": "yellow fever" },{ "name": "yersiniosis" },{ "name": "zika" },{ "name": "legionares" },
-{ "name": "listeriosis" },{ "name": "monkeypox" }, { "name": "COVID-19" }]  }"""
-
-x = json.loads(raw)
-
-ret = []
-
-for item in x['payload']:
-    ret.append(item['name'])
-
-print(ret)
-
-# diseases
-"""
-[ { "name": "unknown" },{ "name": "other" },{ "name": "anthrax cutaneous" },{ "name": "anthrax gastrointestinous" },{ "name": "anthrax inhalation" },{ "name": "botulism" },{ "name": "brucellosis" },{ "name": "chikungunya" }, { "name": "cholera" },
-{ "name": "cryptococcosis" },{ "name": "cryptosporidiosis" },{ "name": "crimean-congo haemorrhagic fever" },{ "name": "dengue" },{ "name": "diphteria" },{ "name": "ebola haemorrhagic fever" },{ "name": "ehec (e.coli)" },{ "name": "enterovirus 71 infection" },{ "name": "influenza a/h5n1" },{ "name": "influenza a/h7n9" },{ "name": "influenza a/h9n2" },{ "name": "influenza a/h1n1" },{ "name": "influenza a/h1n2" },{ "name": "influenza a/h3n5" },{ "name": "influenza a/h3n2" },{ "name": "influenza a/h2n2" },{ "name": "hand, foot and mouth disease" },{ "name": "hantavirus" },{ "name": "hepatitis a" },{ "name": "hepatitis b" },{ "name": "hepatitis c" }, { "name": "hepatitis d" },{ "name": "hepatitis e" },{ "name": "histoplasmosis" },{ "name": "hiv/aids" },{ "name": "lassa fever" },{ "name": "malaria" },{ "name": "marburg virus disease" },{ "name": "measles" },{ "name": "mers-cov" },{ "name": "mumps" },{ "name": "nipah virus" },{ "name": "norovirus infection" },{ "name": "pertussis" },{ "name": "plague" },{ "name": "pneumococcus pneumonia" },{ "name": "poliomyelitis" },{ "name": "q fever" },{ "name": "rabies" },{ "name": "rift valley fever" },{ "name": "rotavirus infection" },{ "name": "rubella" },{ "name": "salmonellosis" },{ "name": "sars" },{ "name": "shigellosis" },{ "name": "smallpox" },{ "name": "staphylococcal enterotoxin b" },{ "name": "thypoid fever" },{ "name": "tuberculosis" },{ "name": "tularemia" },{ "name": "vaccinia and cowpox" },{ "name": "varicella" },{ "name": "west nile virus" },{ "name": "yellow fever" },{ "name": "yersiniosis" },{ "name": "zika" },{ "name": "legionares" },
-{ "name": "listeriosis" },{ "name": "monkeypox" }, { "name": "COVID-19" }] 
-"""
-
-# syndromes
-'''
-[ { "name": "Haemorrhagic Fever" },{ "name": "Acute Flacid Paralysis" },{ "name": "Acute gastroenteritis" },{ "name": "Acute respiratory syndrome" },{ "name": "Influenza-like illness" },{ "name": "Acute fever and rash" },{ "name": "Fever of unknown Origin" },{ "name": "Encephalitis" },{ "name": "Meningitis" }]
-
-'''
