--- PHASE_1/API_SourceCode/scraper/src/scrapy.cfg
@@ -1,11 +1,0 @@
-# Automatically created by: scrapy startproject
-#
-# For more information about the [deploy] section see:
-# https://scrapyd.readthedocs.io/en/latest/deploy.html
-
-[settings]
-default = WHODON_scraper.settings
-
-[deploy]
-#url = http://localhost:6800/
-project = WHODON_scraper
